+++ src/methods/delete_message.rs
@@ -13,7 +13,7 @@ pub struct DeleteMessage<'a> {
     #[serde(skip)]
     proxy: Option<proxy::Proxy>,
     chat_id: types::ChatId<'a>,
-    message_id: u64,
+    message_id: u32,
 }
 
 impl<'a> DeleteMessage<'a> {
@@ -21,7 +21,7 @@ impl<'a> DeleteMessage<'a> {
     pub fn new(
         token: &'a str,
         chat_id: impl Into<types::ChatId<'a>>,
-        message_id: u64,
+        message_id: u32,
     ) -> Self {
         Self {
             token,
