+++ src/methods/send_audio.rs
@@ -12,7 +12,7 @@ pub struct SendAudio<'a> {
     chat_id: types::ChatId<'a>,
     audio: &'a Audio<'a>,
     disable_notification: Option<bool>,
-    reply_to_message_id: Option<u64>,
+    reply_to_message_id: Option<u32>,
     reply_markup: Option<types::raw::Keyboard<'a>>,
 }
 
@@ -41,7 +41,7 @@ impl<'a> SendAudio<'a> {
     }
 
     /// Configures `reply_to_message_id`.
-    pub fn reply_to_message_id(mut self, id: u64) -> Self {
+    pub fn reply_to_message_id(mut self, id: u32) -> Self {
         self.reply_to_message_id = Some(id);
         self
     }
