+++ src/types/file.rs
@@ -8,7 +8,7 @@ pub struct File {
     /// The file's ID.
     pub file_id: String,
     /// The file's size.
-    pub file_size: Option<i64>,
+    pub file_size: Option<u32>,
     /// The file's name.
     pub file_path: Option<String>,
 }
